requirements/base.txt
Django==1.11.2
django-appconf==1.0.2
djangorestframework==3.6.3
django-filter==1.0.4
